application-tier/models/model-chatbot/pyproject.toml
[project]
name = "model-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.0.0",
    "chromadb>=0.4.0",
    "jpype1>=1.4.1",
    "langchain-chroma>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pdfplumber>=0.9.0",
    "pillow>=9.0.0",
    "pypdf2>=3.0.0",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tabula-py>=2.5.0",
    "typing-extensions>=4.0.0",
]
